CPT-111 Enable PDF-A generation after ghostscript update
In commit `84eea7274ffb45a4297634bb3ce636d627897826`, we have updated
ghostscript to fix several vulnerabilities, one of which has been
highlighted in a penetration test.

According to the discussion in the ghostscript bug tracker
(https://bugs.ghostscript.com/show_bug.cgi?id=707450), there was an
unclear specification regarding images with an SMask flag, meaning they
are supposed to be printed in grayscales. The new ghostscript version
returns an error if such an image is detected in a color PDF and does
not convert to PDF/A. In the meantime, there has been a clarification
that this is indeed allowed and the commit has been effectively
reverted. Unfortunately, this fix is not yet officially released.

Therefore, we apply the patch containing the fix to the downloaded
ghostscript version.

Also, we copy a PDF/A definition file and appropriate color profile from
the github source and put them in /etc/odoo to be used for PDF/A
generation.