CPT-127 Make list of attachments accessible again
When opening the list of attachments, Odoo wants to display the total
number of items in the list. This is displayed in the top right such as
"1 - 80 of 2264".

In order to display the total number, Odoo fetches all IDs and filters
out the ids to which the user does not have access. For the JobRad
database with more than 32 million entries this causes either a
MemoryError or a timeout.

The suggested solution is to skip the filtering. This means the total
number of entries will be wrong. The list can still be navigated without
any issues.