CPT-294 Use newer poetry version
Poetry version 2 has been released. It is required for installing the
official Odoo upgrade utils which tremendously help when migrating data.

The parameter `prefer-active-python` has been replaced by the inverse
parameter `use-poetry-python`. Since the default behavior changed and is
now the same as we manually specified before, we do not need to set this
parameter.